Library/requirements.txt
python-dotenv==1.0.0
web3==6.13.0
flask-ngrok==0.0.25